add class PageController, some changes in code, some files have been moved, files main.js has been cleaned
