+++ src/components/amountOfMovies.js
@@ -1,27 +1,16 @@
-import {createElement} from "../utils.js";
+// import {createElement} from "../utils.js";
+import AbstractComponent from "./abstractComponent.js";
 
 export const createAmountOfMovies = (amount) => (`<p>${amount} movies inside</p>`);
 
-export default class AmountOfMoviesFilmComponent {
+export default class AmountOfMoviesFilmComponent extends AbstractComponent {
   constructor(amount) {
+    super();
     this._amount = amount;
-    this._element = null;
   }
 
   getTemplate() {
     return createAmountOfMovies(this._amount);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
