+++ src/components/rankOfUser.js
@@ -1,28 +1,13 @@
-import {createElement} from "../utils.js";
+// import {createElement} from "../utils.js";
+import AbstractComponent from "./abstractComponent.js";
 
 const createRankOfUser = () => (`<section class="header__profile profile">
     <p class="profile__rating">Movie Buff</p>
     <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
   </section>`);
 
-export default class Rank {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Rank extends AbstractComponent {
   getTemplate() {
     return createRankOfUser();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
